refactor: utils 폴더 src 폴더 내부로 이동
