+++ __tests__/utils/gameNumberGenerator.test.js
@@ -1,5 +1,5 @@
 import GameNumberGenerator from '../../src/model/GameNumberGenerator.js';
-import Validators from '../../utils/validator/index.js';
+import Validators from '../../src/utils/validator/index.js';
 
 describe('GameNumberGenerator', () => {
   test('숫자 1~9 범위 내에서 서로 다른 3자리 숫자를 생성한다.  ', () => {
