+++ src/utils/error/GameNumberGeneratorError.js
