+++ src/utils/validator/unit.js
@@ -1,4 +1,4 @@
-import { ERROR_MESSAGE } from '../../src/constants/Messages.js';
+import { ERROR_MESSAGE } from '../../constants/Messages.js';
 import ValidationError from '../error/ValidationError.js';
 
 /**
