+++ src/view/index.js
@@ -1,6 +1,6 @@
 import { OUTPUT_MESSAGES } from '../constants/Messages.js';
 import OutputView from './OutputView.js';
-import Validators from '../../utils/validator/index.js';
+import Validators from '../utils/validator/index.js';
 import InputView from './InputView.js';
 import { HINT } from '../constants/System.js';
